requirements.txt
absl-py==1.1.0
altgraph==0.17.3
astunparse==1.6.3
Bottleneck @ file:///C:/Windows/Temp/abs_3198ca53-903d-42fd-87b4-03e6d03a8381yfwsuve8/croots/recipe/bottleneck_1657175565403/work
cachetools==5.2.0
certifi==2022.12.7
cffi @ file:///D:/bld/cffi_1671179504620/work
charset-normalizer==2.0.12
click==8.1.3
cloudpickle @ file:///tmp/build/80754af9/cloudpickle_1632508026186/work
colorama @ file:///C:/Windows/TEMP/abs_9439aeb1-0254-449a-96f7-33ab5eb17fc8apleb4yn/croots/recipe/colorama_1657009099097/work
cryptography @ file:///D:/bld/cryptography_1669592337584/work
cycler @ file:///tmp/build/80754af9/cycler_1637851556182/work
cytoolz @ file:///C:/b/abs_61m9vzb4qh/croot/cytoolz_1667465938275/work
daal4py==2021.5.0
dask @ file:///C:/ci/dask-core_1658515361249/work
dlib @ file:///D:/bld/dlib_1652035369210/work
dnspython @ file:///home/conda/feedstock_root/build_artifacts/dnspython_1646621171513/work
face-recognition==1.3.0
face-recognition-models==0.3.0
Flask @ file:///C:/b/abs_ef16l83sif/croot/flask_1671217367534/work
Flask-Cors @ file:///tmp/build/80754af9/flask_cors_1614875208036/work
Flask-Login @ file:///C:/b/abs_a82vthl_1_/croot/flask-login_1669052139884/work
Flask-MySQLdb==1.0.1
Flask-SQLAlchemy @ file:///C:/b/abs_0fwc4vuk44/croot/flask-sqlalchemy_1671219142382/work
flatbuffers==1.12
fonttools==4.25.0
fsspec @ file:///C:/b/abs_5bjz6v0w_f/croot/fsspec_1670336608940/work
future==0.18.2
gast==0.4.0
google-auth==2.6.6
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
greenlet @ file:///C:/b/abs_47lk_w2ajq/croot/greenlet_1670513248400/work
grpcio==1.46.3
h5py==3.7.0
idna @ file:///home/conda/feedstock_root/build_artifacts/idna_1663625384323/work
imageio @ file:///C:/Windows/TEMP/abs_24c1b783-7540-4ca9-a1b1-0e8aa8e6ae64hb79ssux/croots/recipe/imageio_1658785038775/work
importlib-metadata==4.11.4
itsdangerous @ file:///tmp/build/80754af9/itsdangerous_1621432558163/work
Jinja2 @ file:///C:/b/abs_7cdis66kl9/croot/jinja2_1666908141852/work
joblib @ file:///tmp/build/80754af9/joblib_1635411271373/work
keras @ file:///C:/Users/builder/adipietro/mc3/tf_env/conda-bld/keras_1661810022369/work/keras-2.9.0-py2.py3-none-any.whl
Keras-Preprocessing @ file:///tmp/build/80754af9/keras-preprocessing_1612283640596/work
kiwisolver @ file:///C:/ci/kiwisolver_1653292408254/work
libclang==14.0.1
locket @ file:///C:/ci/locket_1652904093019/work
mando==0.6.4
Markdown==3.3.7
MarkupSafe @ file:///C:/ci/markupsafe_1654489871526/work
matplotlib @ file:///C:/ci/matplotlib-suite_1647423638658/work
mkl-fft==1.3.1
mkl-random @ file:///C:/ci/mkl_random_1626186184278/work
mkl-service==2.4.0
munkres==1.1.4
mysql-connector-python==8.0.31
mysqlclient==2.1.1
networkx @ file:///C:/ci/networkx_1657698759751/work
numexpr @ file:///C:/b/abs_a7kbak88hk/croot/numexpr_1668713882979/work
numpy==1.22.4
oauthlib==3.2.0
opencv-python==4.6.0.66
opt-einsum==3.3.0
packaging @ file:///tmp/build/80754af9/packaging_1637314298585/work
pandas @ file:///C:/b/abs_c6fuezktfm/croot/pandas_1670425103552/work
partd @ file:///opt/conda/conda-bld/partd_1647245470509/work
pefile==2022.5.30
Pillow==9.0.1
protobuf==4.21.12
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser @ file:///home/conda/feedstock_root/build_artifacts/pycparser_1636257122734/work
pyinstaller==5.7.0
pyinstaller-hooks-contrib==2022.14
pyparsing==3.0.9
python-dateutil @ file:///tmp/build/80754af9/python-dateutil_1626374649649/work
pytube==12.1.2
pytz @ file:///C:/Windows/TEMP/abs_90eacd4e-8eff-491e-b26e-f707eba2cbe1ujvbhqz1/croots/recipe/pytz_1654762631027/work
PyWavelets @ file:///C:/b/abs_a8r4b1511a/croot/pywavelets_1670425185881/work
pywin32-ctypes==0.2.0
PyYAML==6.0
radon==5.1.0
requests==2.27.1
requests-oauthlib==1.3.1
rsa==4.8
scikit-image @ file:///C:/b/abs_63r0vmx78u/croot/scikit-image_1669241746873/work
scikit-learn @ file:///C:/ci/scikit-learn_1642599125707/work
scikit-learn-intelex==2021.20220215.153006
scipy @ file:///C:/ci/scipy_1641555198988/work
seaborn @ file:///C:/b/abs_64ypwhwrqe/croot/seaborn_1669625732106/work
sip==4.19.13
six @ file:///tmp/build/80754af9/six_1644875935023/work
SQLAlchemy @ file:///C:/Windows/Temp/abs_f8661157-660b-49bb-a790-69ab9f3b8f7c8a8s2psb/croots/recipe/sqlalchemy_1657867864564/work
tensorboard==2.9.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.9.1
tensorflow-estimator==2.9.0
tensorflow-io-gcs-filesystem==0.26.0
termcolor==1.1.0
threadpoolctl @ file:///Users/ktietz/demo/mc3/conda-bld/threadpoolctl_1629802263681/work
tifffile==2020.10.1
toolz @ file:///C:/b/abs_cfvk6rc40d/croot/toolz_1667464080130/work
tornado @ file:///C:/ci/tornado_1606942392901/work
tqdm @ file:///C:/b/abs_0axbz66qik/croots/recipe/tqdm_1664392691071/work
typing_extensions==4.2.0
urllib3==1.26.9
Werkzeug @ file:///C:/b/abs_17q5kgb8bo/croot/werkzeug_1671216014857/work
wincertstore==0.2
wrapt==1.14.1
zipp @ file:///C:/ci/zipp_1652274073489/work
